+++ src/app/(content-layout)/articles/_articles/components/Card.tsx
@@ -11,7 +11,7 @@ export default function Card(props: Article) {
   const { title, image, writer, createdAt, likeCount } = props;
 
   return (
-    <div className="border-bg100 bg-bg200 flex h-44 w-full flex-col gap-6 rounded-lg border px-9 py-6">
+    <div className="border-bg100 bg-bg200 flex h-44 w-full flex-col gap-6 rounded-lg border px-7 py-6">
       <Link href={`/articles/${props.id}`}>
         <div className="flex justify-between">
           <h3 className="text-2lg-md text-gray300 max-w-75 truncate">{title}</h3>
@@ -24,7 +24,7 @@ export default function Card(props: Article) {
         </div>
 
         <div className="flex items-center justify-between pt-8">
-          <div className="g-3 flex h-8 w-50 items-center gap-2">
+          <div className="g-3 flex h-8 items-center gap-2">
             <Image
               width={32}
               height={32}
